requirements.txt
alembic==1.12.0
blinker==1.6.2
cachelib==0.10.2
click==8.1.4
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
Pillow==10.1.0
protobuf==4.21.12
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.6
WTForms==3.0.1
